Ajustes a migraciones

+++ database/migrations/2024_10_16_201212_create_team_user_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('team_user', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('team_id');
-            $table->bigInteger('user_id');
+            $table->integer('team_id');
+            $table->integer('user_id');
             $table->string('role')->nullable();
             $table->timestamps();
